refactoring include directory
