+++ src/main.c
@@ -62,7 +62,7 @@ tusb_desc_device_t desc_device =
 //--------------------------------------------------------------------+
 // String Descriptors
 //--------------------------------------------------------------------+
-#include "version.h"
+#include "../include/version.h"
 char const *string_desc_arr[STRING_DESC_ARR_SIZE] =
 {
   (const char[]) { 0x09, 0x04 }, // 0: is supported language is English (0x0409)
